requirements.txt
altgraph==0.17.3
anyio==3.7.1
anytree==2.6.0
appdirs==1.4.4
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.9.3
blinker==1.6.2
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
comtypes==1.2.0
cryptography==41.0.3
docx2pdf==0.1.8
ds-store==1.3.1
dulwich==0.21.5
et-xmlfile==1.1.0
exceptiongroup==1.1.2
fake-useragent==0.1.11
fire==0.5.0
Flask==2.3.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
langdetect==1.0.9
langid==1.1.6
loguru==0.7.0
lxml==4.9.3
mac-alias==2.2.2
MarkupSafe==2.1.3
mysql==0.0.3
mysqlclient==2.2.0
nltk==3.8.1
notion-client==2.0.0
numpy==1.25.1
openpyxl==3.1.2
outcome==1.2.0
pandas==2.0.3
paramiko==3.3.1
passlib==1.7.4
pefile==2023.2.7
Pillow==10.0.0
pyasn1==0.5.0
pycparser==2.21
pyee==8.2.2
pygame==2.5.0
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.5
PyMySQL==0.9.3
PyNaCl==1.5.0
pyppeteer==0.2.5
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-nmap==0.7.1
pytz==2023.3
pywifi==1.1.12
pywin32==306
pywin32-ctypes==0.2.2
regex==2023.8.8
requests==2.23.0
scapy==2.5.0
selenium==4.12.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tenacity==8.2.2
termcolor==2.3.0
threadpool==1.3.2
tqdm==4.65.0
treelib==1.6.4
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
urllib3==2.0.4
websocket-client==1.1.0
websockets==8.1
Werkzeug==2.3.6
win32-setctime==1.1.0
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.1.2
xlutils==2.0.0
xlwt==1.3.0
